handle mapping ref which missed with last commit
